requirements.txt
colorthief==0.2.1
moviepy==1.0.3
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.4.0
progressbar33==2.4
scikit_image==0.19.3
scikit_learn==1.2.1
scipy==1.10.1
skimage==0.0
